Improve webpack chunks
